==> ch_04/1-querying_and_merging.ipynb <==
- query() 메서드로 필터링 쉽게 하기

# Performing Database-style Operations on Dataframes

## About the data
In this notebook, we will using daily weather data that was taken from the [National Centers for Environmental Information (NCEI) API](https://www.ncdc.noaa.gov/cdo-web/webservices/v2). The [`0-weather_data_collection.ipynb`](./0-weather_data_collection.ipynb) notebook contains the process that was followed to collect the data. Consult the dataset's [documentation](https://www1.ncdc.noaa.gov/pub/data/cdo/documentation/GHCND_documentation.pdf) for information on the fields.

*Note: The NCEI is part of the National Oceanic and Atmospheric Administration (NOAA) and, as you can see from the URL for the API, this resource was created when the NCEI was called the NCDC. Should the URL for this resource change in the future, you can search for "NCEI weather API" to find the updated one.*


## Setup

import pandas as pd

weather = pd.read_csv('data/nyc_weather_2018.csv') #긴 형태의 데이터
weather.head()

## Querying DataFrames
## 데이터프레임 질의하기
pandas는 query() 메시지를 지원하므로 부울 마스크를 사용하지 않고도 복잡한 필터를 쉽게 만들 수 있다. 구문은 SQL의 WHERE과 비슷하다.

관측소 ID US1NY(미국 뉴욕)를 포함하는 데이터에서 SNOW 열의 값이 0보다 큰 모든 행의 기상 데이터를 조회한다.

snow_data = weather.query('datatype == "SNOW" and value > 0 and station.str.contains("US1NY")')
snow_data.head()

`weather.db` SQLite 데이터베이스에서 SQL문으로 쓰면 이렇다."

```sql
SELECT * 
FROM weather 
WHERE datatype == "SNOW" AND value > 0 AND station LIKE "%US1NY%"
```

import sqlite3

#DB에서 불러온 데이터와 위에서 query()메서드 쓴 데이터 비교 -> True

with sqlite3.connect('data/weather.db') as connection:
    snow_data_from_db = pd.read_sql(
        'SELECT * FROM weather WHERE datatype == "SNOW" AND value > 0 and station LIKE "%US1NY%"', 
        connection
    )

snow_data.reset_index().drop(columns='index').equals(snow_data_from_db)

부울 마스크로 필터링 한 것과도 똑같다.

weather[
    (weather.datatype == 'SNOW') 
    & (weather.value > 0)
    & weather.station.str.contains('US1NY')
].equals(snow_data)

## Merging DataFrames
## 데이터 병합하기

데이터 병합은 행을 기준으로 정하는 것. (SQL에서 UNION ALL)

병합은 조인(JOIN)이라고 하기도 하는데 아래와 같은 종류가 있다.
- 내부결합(교집합)
- 완전 외부 결합(합집합)
- 왼쪽 결합
- 오른쪽 결합

위에서 눈 데이터를 조회했을 때 똑같이 US1NY가 들어감에도 1월 4일 관측값에 차이가 많이 난다. 아마 관측소의 위치 때문일 것 같다. 고도가 높거나 더 북쪽에 있는 관측소의 눈 관측값이 더 높을 것이다. 관측소가 뉴욕시에서 얼마나 멀리 떨어져 있는지에 따라 코네티컷이나 뉴저지 북부와 같은 다른 곳에서 더 강한 눈보라를 관측했을 수도 있다.

NCEI API stations 종단점은 관측소에 관해 필요한 모든 정보를 제공한다.
(아래 데이터)

station_info = pd.read_csv('data/weather_stations.csv')
station_info.head()

weather 데이터프레임은 아래와 같다.

weather.head()

`station_info.id` 컬럼과 `weather.station` 컬럼을 이용해 매칭해서 조인할 수 있다.

station_info.id.describe()

`station_info` 데이터 프레임이 한 관측소당 하나의 행인 반면, `weather` 데이터 프레임은 한 관측소당 여러 행이 있다. -> 고로 조인을 잘해야 데이터 손실이 없다.

weather.station.describe()

station_info.shape[0], weather.shape[0]

행의 개수를 자주 확인해야 하므로 여러 Dataframe의 행 개수를 반환하는 함수를 만들어놓는 것이 좋다.

*dfs 인수는 함수의 입력을 튜플로 받으며, 행 개수를 계산하고자 입력을 리스트 컴프리헨션으로 반복.

def get_row_count(*dfs):
    return [df.shape[0] for df in dfs]

get_row_count(station_info, weather)

`merge()` 메서드를 사용해 결합하려는 열과 함께 완쪽과 오른쪽 데이터프레임을 지정하면 결합 연산을 할 수 있다. 관측소 ID 열이 이름은 데이터프레임마다 다르므로 left_on과 right_on으로 이름을 지정해야 한다.

- weather은 merge()를 호출하는 데이터 프레임
- station_info는 인수로 전달되는 데이터 프레임
- left_on에는 weather의 station 컬럼이 전달됨.
- right_on에는 station_info의 id 컬럼이 전달됨.

inner_join = weather.merge(station_info, left_on='station', right_on='id')
inner_join.sample(5, random_state=0)

`station` 열과 `id`열의 중복 정보를 제거하려면 결합하기 전에 두 열 중 한 열의 이름을 바꿔 두 열의 이름을 같게 하고 `on` 매개변수에 해당 열의 이름을 명시하면 된다.

weather.merge(station_info.rename(dict(id='station'), axis=1), 
              on='station').sample(5, random_state=0)

station_info 데이터프레임에는 279개의 관측소가 있지만 weather 데이터프레임에는 110개의 관측소만 있다 내부 결합을 하면 기상관측이 없는 모든 관측소는 삭데된다.
결합하면서 열을 남기려면 왼쪽이나 오른쪽 결합을 해야한다.

**왼쪽 결합에서는 남겨둘 열이 있는 데이터프레임을 왼쪽에, 다른 데이터프레임을 오른쪽에 배치해야한다. (오른쪽 결합은 그 반대이다.)**

left_join = station_info.merge(weather, left_on='id', right_on='station', how='left')
right_join = weather.merge(station_info, left_on='station', right_on='id', how='right')

#기상데이터가 없는 관측소는 남아 있지만 모두 null값으로 채워지게 된다.
right_join[right_join.datatype.isna()].head()

station_info를 왼쪽에 두고 왼쪽 결합을 했고, station_info를 오른쪽에 두고 오른쪽 결합을 했으므로 두 결과는 같고 모든 관측소에 대한 기록이 남아있다.

아래는 두 결과가 같다는 것을 증명.

left_join.sort_index(axis=1).sort_values(['date', 'station'], ignore_index=True).equals(
    right_join.sort_index(axis=1).sort_values(['date', 'station'], ignore_index=True)
)

기상 관측값이 없는 모든 관측소의 데이터가 남도록 left join(왼쪽 결합)과 right join(오른쪽 결합)을 했기에 행이 추가되었다. 

get_row_count(inner_join, left_join, right_join)

**outer join**을 하면 모든 데이터가 남는다. 따라서 불일치가 발생하는 정보를 파악할 수 있다.

`indicator = True`를 사용하면 결과 데이터프레임에 각 행이 어느 데이터프레임에서 왔는지 알려주는 열을 추가할 수 있다.

outer_join._merge.unique()

outer_join = weather.merge(
    station_info[station_info.id.str.contains('US1NY')], 
    left_on='station', right_on='id', 
    how='outer', indicator=True
)

#외부 결합의 결과 확인
pd.concat([
    outer_join.query(f'_merge == "{kind}"').sample(2, random_state=0) 
    for kind in outer_join._merge.unique()
]).sort_index()

These joins are equivalent to their SQL counterparts. Below is the inner join. Note that to use `equals()` you will have to do some manipulation of the dataframes to line them up:

import sqlite3

with sqlite3.connect('data/weather.db') as connection:
    inner_join_from_db = pd.read_sql(
        'SELECT * FROM weather JOIN stations ON weather.station == stations.id', 
        connection
    )

inner_join_from_db.shape == inner_join.shape

Revisiting the dirty data from chapter 3's [`5-handling_data_issues.ipynb`](../ch_03/5-handling_data_issues.ipynb) notebook.

Data meanings:
- `PRCP`: precipitation in millimeters
- `SNOW`: snowfall in millimeters
- `SNWD`: snow depth in millimeters
- `TMAX`: maximum daily temperature in Celsius
- `TMIN`: minimum daily temperature in Celsius
- `TOBS`: temperature at time of observation in Celsius
- `WESF`: water equivalent of snow in millimeters


Read in the data, dropping duplicates and the uninformative `SNWD` column:

dirty_data = pd.read_csv(
    'data/dirty_data.csv', index_col='date'
).drop_duplicates().drop(columns='SNWD')
dirty_data.head()

We need to create two dataframes for the join. We will drop some unecessary columns as well for easier viewing:

valid_station = dirty_data.query('station != "?"').drop(columns=['WESF', 'station'])
station_with_wesf = dirty_data.query('station == "?"').drop(columns=['station', 'TOBS', 'TMIN', 'TMAX'])

Our column for the join is the index in both dataframes, so we must specify `left_index` and `right_index`:

valid_station.merge(
    station_with_wesf, how='left', left_index=True, right_index=True
).query('WESF > 0').head()

The columns that existed in both dataframes, but didn't form part of the join got suffixes added to their names: `_x` for columns from the left dataframe and `_y` for columns from the right dataframe. We can customize this with the `suffixes` argument:

valid_station.merge(
    station_with_wesf, how='left', left_index=True, right_index=True, suffixes=('', '_?')
).query('WESF > 0').head()

Since we are joining on the index, an easier way is to use the `join()` method instead of `merge()`. Note that the suffix parameter is now `lsuffix` for the left dataframe's suffix and `rsuffix` for the right one's:

valid_station.join(station_with_wesf, how='left', rsuffix='_?').query('WESF > 0').head()

Joins can be very resource-intensive, so it's a good idea to figure out what type of join you need using set operations before trying the join itself. The `pandas` set operations are performed on the index, so whichever columns we will be joining on will need to be the index. Let's go back to the `weather` and `station_info` dataframes and set the station ID columns as the index:

weather.set_index('station', inplace=True)
station_info.set_index('id', inplace=True)

The intersection will tell us the stations that are present in both dataframes. The result will be the index when performing an inner join:

weather.index.intersection(station_info.index)

The set difference will tell us what we lose from each side. When performing an inner join, we lose nothing from the `weather` dataframe:

weather.index.difference(station_info.index)

We lose 169 stations from the `station_info` dataframe, however:

station_info.index.difference(weather.index)

The symmetric difference tells us what we lose from both sides. It is the combination of the set differences in each direction:

ny_in_name = station_info[station_info.index.str.contains('US1NY')]

ny_in_name.index.difference(weather.index).shape[0]\
+ weather.index.difference(ny_in_name.index).shape[0]\
== weather.index.symmetric_difference(ny_in_name.index).shape[0]

The union will show us everything that will be present after a full outer join. Note that we pass in the unique values of the index to make sure we can see the number of stations we will be left with:

weather.index.unique().union(station_info.index)

Note that the symmetric difference is actually the union of the set differences:

ny_in_name = station_info[station_info.index.str.contains('US1NY')]

ny_in_name.index.difference(weather.index).union(weather.index.difference(ny_in_name.index)).equals(
    weather.index.symmetric_difference(ny_in_name.index)
)

<hr>
<div>
    <a href="../ch_03/5-handling_data_issues.ipynb">
        <button>&#8592; Chapter 3</button>
    </a>
    <a href="./0-weather_data_collection.ipynb">
        <button>Weather Data Collection</button>
    </a>
    <a href="./2-dataframe_operations.ipynb">
        <button style="float: right;">Next Notebook &#8594;</button>
    </a>
</div>
<hr>
